requirements.txt
# Core web framework
flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7  # Flask dependency
gunicorn==21.2.0  # For production deployment

# Data processing and analysis
pandas==2.1.1
numpy==1.26.0
networkx==3.1

# Visualization
plotly==5.17.0

# Excel file handling
xlrd>=2.0.1  # For Excel .xls support
openpyxl>=3.1.2  # For Excel .xlsx files

# NLP and embeddings processing
scikit-learn==1.3.1  # For embeddings processing and ML operations
nltk==3.8.1  # For text processing

# Configuration and utilities
python-dotenv==1.0.0
requests==2.31.0  # For HTTP requests
jsonschema==4.19.1  # For JSON validation

# Database
psycopg2-binary>=2.9.3